requirements.txt
flask==3.0.0
python_dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
flask-smorest==0.42.3
flask_jwt_extended==4.6.0
flask_cors==4.0.0
passlib==1.7.4
waitress==2.1.2
psycopg2-binary==2.9.9
flask_login==0.6.3
Flask-Migrate==4.0.5
requests==2.31.0
plotly==5.20.0
dash==2.16.1
kaleido==0.2.1